requirements-dev.txt
pytest
pytest-cov
requests-mock
black
isort
flake8
pre-commit
mypy
build
twine